Update coreutil.py
Fixes #6

+++ src/helpers/coreutil.py
@@ -8,6 +8,7 @@ import re
 import zipfile
 import io
 import shutil
+import json
 
 class Validator(object):
     @staticmethod
@@ -36,12 +37,18 @@ class PresenceVerifier(object):
             return True
         else:
             if auto_install:
-                print('Donwloading and extracting DocFX...')
-                urllib.request.urlretrieve('https://github.com/dotnet/docfx/releases/download/v2.42.4/docfx.zip', 'temp_docfx.zip')
-                with zipfile.ZipFile('temp_docfx.zip', 'r') as zip_ref:
-                    zip_ref.extractall('dbin/docfx')
-                os.remove('temp_docfx.zip')
-                return True
+                print('[info] Downloading and extracting DocFX...')
+                docfx_url = GitHubUtil.get_latest_release('dotnet/docfx')
+
+                if docfx_url:
+                    print(f'[info] Found latest DocFX at {docfx_url}')
+                    urllib.request.urlretrieve(docfx_url, 'temp_docfx.zip')
+                    with zipfile.ZipFile('temp_docfx.zip', 'r') as zip_ref:
+                        zip_ref.extractall('dbin/docfx')
+                    os.remove('temp_docfx.zip')
+                    return True
+                
+                print('[info] Could not download DocFX.')
             return False
 
     @staticmethod
@@ -104,3 +111,19 @@ class ConsoleUtil(object):
         output = output.split('\\n')
         for x in range(len(output)):
             print (output[x])
+
+class GitHubUtil(object):
+    @staticmethod
+    def get_latest_release(repository):
+        url = f'https://api.github.com/repos/{repository}/releases/latest'
+        request = urllib.request.Request(url)
+
+        response = urllib.request.urlopen(request).read()
+        content = json.loads(response.decode('utf-8'))
+
+        asset_container = content['assets']
+        target_url = next((item for item in asset_container if item['browser_download_url'].endswith('/docfx.zip')), None)
+
+        if target_url:
+            return target_url['browser_download_url']
+        return target_url
